add data transmission functions in DataTrans Class
